fix: reorganize server files for admin dashboard
